+++ plugins/org.eclipse.ocl/src/org/eclipse/ocl/cst/impl/IntegerLiteralExpCSImpl.java
@@ -13,10 +13,11 @@
  *
  * </copyright>
  *
- * $Id: IntegerLiteralExpCSImpl.java,v 1.3 2008/11/30 22:11:37 cdamus Exp $
+ * $Id: IntegerLiteralExpCSImpl.java,v 1.3.10.1 2010/01/15 17:27:37 ewillink Exp $
  */
 package org.eclipse.ocl.cst.impl;
 
+import java.math.BigInteger;
 import org.eclipse.emf.common.notify.Notification;
 import org.eclipse.emf.ecore.EClass;
 import org.eclipse.emf.ecore.impl.ENotificationImpl;
@@ -48,7 +49,7 @@ public class IntegerLiteralExpCSImpl
 	 * @generated
 	 * @ordered
 	 */
-	protected static final Integer INTEGER_SYMBOL_EDEFAULT = null;
+	protected static final BigInteger INTEGER_SYMBOL_EDEFAULT = null;
 
 	/**
 	 * The cached value of the '{@link #getIntegerSymbol() <em>Integer Symbol</em>}' attribute.
@@ -58,7 +59,7 @@ public class IntegerLiteralExpCSImpl
 	 * @generated
 	 * @ordered
 	 */
-	protected Integer integerSymbol = INTEGER_SYMBOL_EDEFAULT;
+	protected BigInteger integerSymbol = INTEGER_SYMBOL_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -81,20 +82,22 @@ public class IntegerLiteralExpCSImpl
 
 	/**
 	 * <!-- begin-user-doc -->
+	 * @since 3.0
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public Integer getIntegerSymbol() {
+	public BigInteger getIntegerSymbol() {
 		return integerSymbol;
 	}
 
 	/**
 	 * <!-- begin-user-doc -->
+	 * @since 3.0
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setIntegerSymbol(Integer newIntegerSymbol) {
-		Integer oldIntegerSymbol = integerSymbol;
+	public void setIntegerSymbol(BigInteger newIntegerSymbol) {
+		BigInteger oldIntegerSymbol = integerSymbol;
 		integerSymbol = newIntegerSymbol;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET,
@@ -125,7 +128,7 @@ public class IntegerLiteralExpCSImpl
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case CSTPackage.INTEGER_LITERAL_EXP_CS__INTEGER_SYMBOL :
-				setIntegerSymbol((Integer) newValue);
+				setIntegerSymbol((BigInteger) newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
